chore: move time functions to utils package
