+++ pkg/utils/time.go
@@ -1,4 +1,4 @@
-package iovns
+package utils
 
 import (
 	"time"
