+++ x/starname/controllers/account/account.go
@@ -11,7 +11,6 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/iov-one/iovns"
 	"github.com/iov-one/iovns/x/configuration"
 	"github.com/iov-one/iovns/x/starname/keeper"
 	"github.com/iov-one/iovns/x/starname/types"
@@ -162,7 +161,7 @@ func (a *Account) notExpired() error {
 		return nil
 	}
 	// check if account has expired
-	expireTime := iovns.SecondsToTime(a.account.ValidUntil)
+	expireTime := utils.SecondsToTime(a.account.ValidUntil)
 	if !expireTime.Before(a.ctx.BlockTime()) {
 		return nil
 	}
@@ -181,7 +180,7 @@ func (a *Account) renewable() error {
 	a.requireConfiguration()
 
 	// do calculations
-	newValidUntil := iovns.SecondsToTime(a.account.ValidUntil).Add(a.conf.AccountRenewalPeriod)
+	newValidUntil := utils.SecondsToTime(a.account.ValidUntil).Add(a.conf.AccountRenewalPeriod)
 	// renew count bumped because domain is already at count 1 when created
 	renewCount := a.conf.AccountRenewalCountMax + 1
 	// set new expected valid until
@@ -431,7 +430,7 @@ func (a *Account) gracePeriodFinished() error {
 	}
 	// get grace period and expiration time
 	gracePeriod := a.conf.AccountGracePeriod
-	expireTime := iovns.SecondsToTime(a.account.ValidUntil)
+	expireTime := utils.SecondsToTime(a.account.ValidUntil)
 	if a.ctx.BlockTime().After(expireTime.Add(gracePeriod)) {
 		return nil
 	}
